feat: 필요한 response 필드 추가

+++ backend/src/main/java/com/happy/friendogly/playground/dto/response/FindPlaygroundDetailResponse.java
@@ -7,8 +7,8 @@ import java.util.List;
 public record FindPlaygroundDetailResponse(
 
         Long id,
-        long totalPetCount,
-        long participatePetCount,
+        int totalPetCount,
+        int participatePetCount,
         List<PlaygroundPetDetail> playgroundPetDetails
 ) {
